Title: 정수 사각형 최대 합, Time: 11ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240717/정수 사각형 최대 합/maximum-sum-path-in-square.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 #define MX 105
-int a[MX][MX], d[MX][MX];
+long long a[MX][MX], d[MX][MX];
 
 int main() {
     int n;
